requirements.txt
asgiref==3.8.1
charset-normalizer==3.4.2
Django==5.2.1
django-pwa==2.0.1
djangorestframework==3.16.0
pillow==11.2.1
reportlab==4.4.2
sqlparse==0.5.3
tzdata==2025.2
